Update protobuf GetSchema contract to include key column information. Closes #57
